Encapsulate scr-SDL interface into a separate file
