+++ Makefile
@@ -47,11 +47,11 @@ UTILS = maketape readtape
 #
 
 SRCS =	access.c boot.c branch.c conf.c covox.c double.c ea.c itab.c \
-	main.c service.c ui.c scr.c timer.c tape.c disk.c mouse.c printer.c \
+	main.c service.c ui.c scr-sdl.c scr.c timer.c tape.c disk.c mouse.c printer.c \
 	single.c weird.c tty.c io.c timing.c sound.c disas.c serial.c bkplip.c \
 	terakdisk.c synth.c emu2149.c
 OBJS =	access.o boot.o branch.o conf.o covox.o double.o ea.o itab.o icon.o \
-	main.o service.o ui.o scr.o timer.o tape.o disk.o mouse.o printer.o \
+	main.o service.o ui.o scr-sdl.o scr.o timer.o tape.o disk.o mouse.o printer.o \
 	single.o weird.o tty.o io.o timing.o sound.o disas.o serial.o bkplip.o \
 	terakdisk.o synth.o emu2149.o
 INCS =	defines.h scr.h conf.h emu2149.h emutypes.h
