+++ scr.h
@@ -3,18 +3,5 @@
  * User-modifiable parameters of the video driver.
  */
 
-#define Black {0,0,0}
-#define Red {255,0,0}
-#define Green {0,255,0}
-#define Blue {0,0,255}
-#define Yellow {255,255,0}
-#define Cyan {0,255,255}
-#define Magenta {255,0,255}
-#define Brown {128,0,0}
-#define Violet {192,0,192}
-#define DarkRed {192,0,0}
-#define LightBlue {128,0,255}
-#define White {255,255,255}
-
 extern int horsize, vertsize;
 extern void Quit_SDL(void);
